Extracted normal input
