+++ lib/views/login_view.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_ecoapp/models/store.dart';
-import 'package:flutter_ecoapp/views/debug/debug.dart';
 import 'package:flutter_ecoapp/views/style/colors.dart';
 import 'package:flutter_ecoapp/views/widgets/bottom_nav_bar.dart';
 import 'package:flutter_ecoapp/views/widgets/google_button.dart';
 import 'package:flutter_ecoapp/views/widgets/normal_button.dart';
-import 'package:flutter_ecoapp/views/widgets/search_bar.dart';
-import 'package:flutter_ecoapp/views/widgets/store/storeview/store_cover_section.dart';
+import 'package:flutter_ecoapp/views/widgets/normal_input.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class LoginView extends StatelessWidget {
@@ -52,16 +49,8 @@ class LoginView extends StatelessWidget {
         ),
         child: Column(
           children: [
-            getInput(
-              header: 'Email',
-              hint: 'Ingresa tu email',
-              icon: Icons.mail
-            ),
-            getInput(
-              header: 'Contraseña',
-              hint: 'Ingresa tu contraseña',
-              icon: Icons.vpn_key
-            ),
+            NormalInput(header: 'Email', hint: 'Ingresa tu email', icon: Icons.mail),
+            NormalInput(header: 'Contraseña', hint: 'Ingresa tu contraseña', icon: Icons.vpn_key),
             Container(
               margin: EdgeInsets.symmetric(
                 horizontal: 40.0
@@ -98,30 +87,7 @@ class LoginView extends StatelessWidget {
     );
   }
 
-  Widget getInput({String header, String hint, IconData icon}){
-    return Container(
-      margin: EdgeInsets.only(
-        bottom: 20.0
-      ),
-      child: TextField(
-        style: GoogleFonts.montserrat(),
-        decoration: InputDecoration(
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(20.0)
-          ),
-          hintText: hint,
-          contentPadding: EdgeInsets.symmetric(
-            horizontal: 20.0,
-            vertical: 5.0
-          ),
-          floatingLabelBehavior: FloatingLabelBehavior.auto,
-          labelText: header,
-          prefixIcon: Icon(icon)
-        ),
-      ),
-    );
-  }
-
 }
+
