+++ lib/views/widgets/normal_input.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class NormalInput extends StatelessWidget {
+  const NormalInput({
+    Key key,
+    @required this.header,
+    @required this.hint,
+    @required this.icon,
+  }) : super(key: key);
+
+  final String header;
+  final String hint;
+  final IconData icon;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsets.only(
+        bottom: 20.0
+      ),
+      child: TextField(
+        style: GoogleFonts.montserrat(),
+        decoration: InputDecoration(
+          border: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(20.0)
+          ),
+          hintText: hint,
+          contentPadding: EdgeInsets.symmetric(
+            horizontal: 20.0,
+            vertical: 5.0
+          ),
+          floatingLabelBehavior: FloatingLabelBehavior.auto,
+          labelText: header,
+          prefixIcon: Icon(icon)
+        ),
+      ),
+    );
+  }
+}
